requirements.txt
numpy
scipy
easydict
pyyaml